safari-600.7-branch/LayoutTests/fast/dom/shadow/get-distributed-nodes-orphan-expected.txt

getDistributedNodes() should work out of Document

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


getDistributedNodes() should work out of Document
PASS nodeList.length is expectedNodes.length
PASS nodeList.item(0) is expectedNodes[0]

When a content element is disconnected from ShadowRoot, it should not work.
PASS nodeList.length is expectedNodes.length

Reprojection case
PASS nodeList.length is expectedNodes.length
PASS nodeList.item(0) is expectedNodes[0]
PASS nodeList.length is expectedNodes.length
PASS nodeList.item(0) is expectedNodes[0]
PASS nodeList.item(1) is expectedNodes[1]

rootChild is disconnected. Now content became inactive, so content element itself should be distributed.
PASS nodeList.length is expectedNodes.length
PASS nodeList.length is expectedNodes.length
PASS nodeList.item(0) is expectedNodes[0]
PASS nodeList.item(1) is expectedNodes[1]

PASS successfullyParsed is true

TEST COMPLETE
